// web/models/datasets.ts
export enum DataSourceType {
    FILE = 'upload_file',
    NOTION = 'notion_import',
    WEB = 'web_import',
    DRIVE = 'drive_import'
}

export type DataSet = {
    id: string
    name: string
    icon: string
    icon_background: string
    description: string
    permission: 'only_me' | 'all_team_members'
    data_source_type: DataSourceType
    indexing_technique: 'high_quality' | 'economy'
    created_by: string
    updated_by: string
    updated_at: number
    app_count: number
    document_count: number
    word_count: number
    embedding_model: string
    embedding_model_provider: string
    embedding_available: boolean
    retrieval_model_dict: RetrievalConfig
    retrieval_model: RetrievalConfig
    tags: Tag[]
}

export type CustomFile = File & {
    id?: string
    extension?: string
    mime_type?: string
    created_by?: string
    created_at?: number
}

export type FileItem = {
    fileID: string
    file: CustomFile
    progress: number
}
// ...
export type DocumentIndexingStatus = typeof DocumentIndexingStatusList[number]

export const DisplayStatusList = [
    'queuing',
    'indexing',
    'paused',
    'error',
    'available',
    'enabled',
    'disabled',
    'archived',
] as const

export type DocumentDisplayStatus = typeof DisplayStatusList[number]

export type DataSourceInfo = {
    upload_file: {
        id: string
        name: string
        size: number
        mime_type: string
        created_at: number
        created_by: string
        extension: string
    }
    notion_page_icon?: string
}

export type InitialDocumentDetail = {
    id: string
    batch: string
    position: number
    dataset_id: string
    data_source_type: DataSourceType
    data_source_info: DataSourceInfo
    dataset_process_rule_id: string
    name: string
    created_from: 'api' | 'web'
    created_by: string
    created_at: number
    indexing_status: DocumentIndexingStatus
    display_status: DocumentDisplayStatus
    completed_segments?: number
    total_segments?: number
    doc_form: 'text_model' | 'qa_model'
}
// ...
export type createDocumentResponse = {
    dataset?: DataSet
    batch: string
    documents: InitialDocumentDetail[]
}